--- packages/data/galaxy/schema
@@ -1,1 +1,0 @@
-../../../lib/galaxy/schema
