+++ packages/schema/HISTORY.rst
@@ -1,0 +1,10 @@
+History
+-------
+
+.. to_doc
+
+---------------------
+21.1.0.dev0
+---------------------
+
+* First release.
