+++ packages/schema/MANIFEST.in
@@ -1,0 +1,1 @@
+include *.rst *.txt LICENSE */py.typed
