+++ packages/schema/galaxy/schema
@@ -1,0 +1,1 @@
+../../../lib/galaxy/schema/
