+++ packages/schema/mypy.ini
@@ -1,0 +1,1 @@
+../../mypy.ini
