+++ packages/schema/pyproject.toml
@@ -1,0 +1,1 @@
+../package-pyproject.toml
