+++ packages/schema/scripts
@@ -1,0 +1,1 @@
+../build_scripts
