+++ packages/schema/tests/schema
@@ -1,0 +1,1 @@
+../../../test/unit/schema/
