+++ packages/tool_shed/LICENSE
@@ -1,0 +1,1 @@
+../../LICENSE.txt
