+++ packages/tool_shed/Makefile
@@ -1,0 +1,1 @@
+../package.Makefile
