+++ packages/tool_shed/dev-requirements.txt
@@ -1,0 +1,1 @@
+../package-dev-requirements.txt
