+++ packages/tool_shed/setup.cfg
@@ -1,0 +1,40 @@
+[metadata]
+author = Galaxy Project and Community
+author_email = galaxy-committers@lists.galaxyproject.org
+classifiers =
+    Development Status :: 5 - Production/Stable
+    Environment :: Console
+    Intended Audience :: Developers
+    License :: OSI Approved :: Academic Free License (AFL)
+    Natural Language :: English
+    Operating System :: POSIX
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Topic :: Software Development
+    Topic :: Software Development :: Code Generators
+    Topic :: Software Development :: Testing
+description = Galaxy auth framework and implementations
+keywords =
+    Galaxy
+license = AFL
+license_files =
+    LICENSE
+long_description = file: README.rst, HISTORY.rst
+long_description_content_type = text/x-rst
+name = galaxy-tool-shed
+url = https://github.com/galaxyproject/galaxy
+version = 23.1.0.dev0
+
+[options]
+include_package_data = True
+install_requires =
+    galaxy-webapps
+packages = find:
+python_requires = >=3.7
+
+[options.packages.find]
+exclude =
+    tests*
