+++ packages/tool_shed/tests/tool_shed
@@ -1,0 +1,1 @@
+../../../test/unit/tool_shed
