+++ test/unit/tool_shed/test_installed_repository_manager.py
