+++ test/unit/tool_shed/test_tool_panel_manager.py
